Added a featured image and a lead text to page search list cards
PT-1126.
