+++ src/headless-cms/components/CmsPageSearch/cmspagesearch.module.scss
@@ -1,0 +1,18 @@
+.searchForm {
+  margin-bottom: 30px;
+}
+
+.card {
+  > div {
+    margin-left: 120px;
+  }
+  margin: 10px 0 20px 0;
+  cursor: pointer;
+}
+
+.cardImage {
+  width: 100px;
+  height: 100px;
+  margin-top: -110px;
+  display: block;
+}
